bioguard-ai-service/requirements.txt
# FastAPI Server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
jinja2==3.1.4

# AI/ML
onnxruntime>=1.20.0,<2.0.0
numpy==1.26.4
opencv-python-headless==4.10.0.84
torch>=2.2.0,<3.0.0

# Image Processing
Pillow==10.4.0

# Utilities
python-dotenv==1.0.1
pydantic==2.9.0

easydict
numpy
tqdm
torchvision
torch
opencv_python
Pillow
tensorboardX
fastapi
uvicorn[standard]
python-multipart
jinja2
aiofiles
onnxruntime
